f2fs_io: Fix integer multiplication overflow error in fiemap
When using fiemap to obtain the block address of files
larger than 2GB ((2147483647+1) bytes), an integer
multiplication overflow error will occur.
This issue is caused by the following code:
    start = atoi(argv[1]) * F2FS_BLKSIZE;
    length = atoi(argv[2]) * F2FS_BLKSIZE;

+++ tools/f2fs_io/f2fs_io.c
@@ -809,8 +809,8 @@ static void do_fiemap(int argc, char **argv, const struct cmd_desc *cmd)
 	}
 
 	memset(fm, 0, sizeof(struct fiemap));
-	start = atoi(argv[1]) * F2FS_BLKSIZE;
-	length = atoi(argv[2]) * F2FS_BLKSIZE;
+	start = (u64)atoi(argv[1]) * F2FS_BLKSIZE;
+	length = (u64)atoi(argv[2]) * F2FS_BLKSIZE;
 	fm->fm_start = start;
 	fm->fm_length = length;
